fix(main/ffmpeg): fix build with ndk-r25
Configure with --disable-vulkan to avoid error due to vulkan.h trying
to include vulkan_beta.h from a third_party directory in the ndk.
Remove -fno-integrated-as flag which no longer works with ndk-r25 as
the non-internal (non-llvm) assembler is not included in the ndk anymore.

Also disable vpx and x265 support, libvpx and libx265 has compilation
issues (cannot compile with llvm assembler).  Disable until we have
managed to fix that.